Cleaned up the bundle enhancing methods by moving them out into their own class
